Connected upload endpoint with FileDetailForm and renamed DocumentData Folder in API to "Data". Also added appsettings.json file
