+++ API/appsettings.Development.json
@@ -1,9 +1,5 @@
 {
-  "ConnectionStrings": {
-    "DefaultConnection": "server=localhost;database=modulo;trusted_connection=true",
-    "AzureConnection": "DefaultEndpointsProtocol=https;AccountName=binuliazurestore;AccountKey=nQOenPLhON2FgwgZ3BIGwUsZjzp1dwD+Yf2sI7tEFdnkXXPyGVf7IL6/Nj4RjJswRrckdNg5z4EA+ASt5AxJYQ==;EndpointSuffix=core.windows.net"
   
-},
 
   "Logging": {
     "LogLevel": {
