+++ client/src/components/DocumentManager/Document-uplaod/FileSelector.tsx
@@ -1,4 +1,4 @@
-import React, { Dispatch, SetStateAction, useRef, useState } from "react";
+import React, { Dispatch, SetStateAction, useState } from "react";
 
 import { Documents } from "./Documents";
 import FileDetailForm from "./FileDetailForm";
